auth+middleware first commit
