+++ views/admin/showmahasiswa.pug
@@ -1,0 +1,30 @@
+extends ../layout/master
+block title
+    title Daftar Mahasiswa
+block custom-css
+    link(href="vendor/datatables/dataTables.bootstrap4.min.css" rel="stylesheet")
+block web-content
+    .card.shadow.mb-4
+        .card-header.py-3
+            h6.m-0.font-weight-bold.text-primary Daftar Mahasiswa
+        .card-body
+            .table-responsive
+                table.table.table-bordered(id="dataTable", width="100%", cellspacing="0")
+                    thead
+                        tr
+                            th NRP
+                            th User ID
+                            th Nama Mahasiswa
+                            th Program Studi ID
+                            th IPK Terakhir
+                            th Status Aktif
+                    tbody
+                        each m in mahasiswa
+                            tr
+                                td= m.nrp
+                                td= m.user_id
+                                td= m.nama_mahasiswa
+                                td= m.prodi
+                                td= m.ipk
+                                td= m.status
+
