+++ views/auth/login.pug
@@ -1,0 +1,65 @@
+doctype html
+html(lang="en")
+    head
+        meta(charset="utf-8")
+        meta(http-equiv="X-UA-Compatible" content="IE=edge")
+        meta(name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no")
+        meta(name="description" content="")
+        meta(name="author" content="")
+
+        title Login
+
+        // Custom fonts for this template
+        link(href="/vendor/fontawesome-free/css/all.min.css" rel="stylesheet" type="text/css")
+        link(href="https://fonts.googleapis.com/css?family=Nunito:200,200i,300,300i,400,400i,600,600i,700,700i,800,800i,900,900i" rel="stylesheet")
+
+        // Custom styles for this template
+        link(href="/css/sb-admin-2.min.css" rel="stylesheet")
+
+    body(class="bg-gradient-primary")
+        .container
+            // Outer Row
+            .row.justify-content-center
+                .col-xl-10.col-lg-12.col-md-9
+                    .card.o-hidden.border-0.shadow-lg.my-5
+                        .card-body.p-0
+                            // Nested Row within Card Body
+                            .row
+                                .col-lg-6.d-none.d-lg-block.bg-login-image
+                                .col-lg-6
+                                    .p-5
+                                        .text-center
+                                            h1.h4.text-gray-900.mb-4 Welcome Back!
+                                        form.user(method="post" action="/auth/login")
+                                            .form-group
+                                                input.form-control.form-control-user(type="text" name="username" placeholder="Enter Username...")
+                                            .form-group
+                                                input.form-control.form-control-user(type="password" name="password" placeholder="Password")
+                                            .form-group
+                                                .custom-control.custom-checkbox.small
+                                                    input.custom-control-input(type="checkbox" id="customCheck")
+                                                    label.custom-control-label(for="customCheck") Remember Me
+                                            button.btn.btn-primary.btn-user.btn-block(type="submit") Login
+                                            hr
+                                            a.btn.btn-google.btn-user.btn-block(href="/auth/login")
+                                                i.fab.fa-google.fa-fw
+                                                |  Login with Google
+                                            a.btn.btn-facebook.btn-user.btn-block(href="/auth/login")
+                                                i.fab.fa-facebook-f.fa-fw
+                                                |  Login with Facebook
+
+                                        hr
+                                        .text-center
+                                            a.small(href="forgot-password.html") Forgot Password?
+                                        .text-center
+                                            a.small(href="register.html") Create an Account!
+
+        // Bootstrap core JavaScript
+        script(src="/vendor/jquery/jquery.min.js")
+        script(src="/vendor/bootstrap/js/bootstrap.bundle.min.js")
+
+        // Core plugin JavaScript
+        script(src="/vendor/jquery-easing/jquery.easing.min.js")
+
+        // Custom scripts for all pages
+        script(src="/js/sb-admin-2.min.js")
