+++ views/mahasiswa/index.pug
@@ -1,4 +1,6 @@
 extends ../layout/master
 
+
+
 block web-content
     h1= message
