--- views/mahasiswa/showmahasiswa.pug
@@ -1,22 +1,0 @@
-extends ../layout/master
-
-block web-content
-    h1 Daftar Mahasiswa
-    table(border="1")
-        thead
-            tr
-                th NRP
-                th User ID
-                th Nama Mahasiswa
-                th Program Studi ID
-                th IPK Terakhir
-                th Status Aktif
-        tbody
-            each m in mahasiswa
-                tr
-                    td= m.nrp
-                    td= m.user_id
-                    td= m.nama_mahasiswa
-                    td= m.prodi
-                    td= m.ipk
-                    td= m.status
